+++ CHANGELOG.md
@@ -152,3 +152,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Provide also human readable by fetch all registered bridgeheads
 - DataSHIELD Project Status Error
 - Catch token manager exceptions and return ERROR status
+- Send emails in async execution pool