+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.1 - 2024-11-14]
+## [0.0.1 - 2024-11-15]
 ### Added
 - First version of the project
 - Spring Application
@@ -155,3 +155,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Send emails in async execution pool
 - Notification async execution pool
 - Exporter async execution pool
+- Action explanations